refactor: :zap: time picker with the implement responsive device detection and theme management hooks
