+++ apps/zenithui-docs/src/app/global.css
@@ -140,12 +140,3 @@
     @apply bg-background text-foreground;
   } */
 }
-
-@layer base {
-  * {
-    @apply border-border outline-ring/50;
-  }
-  body {
-    @apply bg-background text-foreground;
-  }
-}
