+++ packages/zenithui-time-picker/src/components/toggle-group/index.tsx
@@ -1,7 +1,5 @@
-"use-client"
-
 import * as React from "react"
-import { cn } from "../../utils"
+import { cn } from "../utils"
 
 type ToggleGroupType = "single" | "multiple"
