+++ packages/zenithui-time-picker/src/index.ts
@@ -1,3 +1,3 @@
-export { TimePicker } from "./time-picker"
-export { default as CountDownTimer } from "./count-down"
 import "./index.css"
+export { default as TimePicker } from "./components/time-picker"
+export { default as CountDownTimer } from "./components/count-down"
